Fix palatalization
The list modifications that were being made in the un_palatalize function were
being ignored.

I've also removed the stem_length function because it doesn't make much sense
and isn't very useful in the light of the fact that the stem of the word isn't
neceserily just a mere part of it (see the un_palatalize function).